challenges/obfuscation_drive/src/website/requirements.txt
click==8.1.3
commonmark==0.9.1
Deprecated==1.2.13
Flask==2.2.2
Flask-Limiter==2.7.0
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==2.7.1
MarkupSafe==2.1.1
packaging==21.3
pycryptodome==3.15.0
Pygments==2.13.0
pyparsing==3.0.9
rich==12.6.0
typing_extensions==4.4.0
Werkzeug==2.2.2
wrapt==1.14.1
